rename: dto를 ui 패키지로 이동
