+++ src/main/java/wooteco/subway/service/StationService.java
@@ -9,8 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.StationRequest;
-import wooteco.subway.dto.StationResponse;
+import wooteco.subway.ui.dto.StationRequest;
+import wooteco.subway.ui.dto.StationResponse;
 import wooteco.subway.exception.EmptyResultException;
 
 @Service
