+++ src/main/java/wooteco/subway/ui/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.ui.dto;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
