+++ src/main/java/wooteco/subway/ui/dto/PathResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.ui.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
